% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_importance_mean.R
\name{extract_importance_mean}
\alias{extract_importance_mean}
\title{Extract the learner-specific importance from a mean object}
\usage{
extract_importance_mean(fit, feature_names, coef = 0)
}
\arguments{
\item{fit}{the \code{mean} object.}

\item{feature_names}{the feature names}

\item{coef}{the Super Learner coefficient associated with the learner.}
}
\value{
a tibble, with columns \code{algorithm} (the fitted algorithm),
\code{feature} (the feature), \code{importance} (the algorithm-specific
extrinsic importance of the feature), \code{rank} (the feature importance
rank, with 1 indicating the most important feature), and \code{weight}
(the algorithm's weight in the Super Learner)
}
\description{
Extract the individual-algorithm extrinsic importance from a mean object,
along with the importance rank.
}
