% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_set.R
\name{get_base_set}
\alias{get_base_set}
\title{Get an initial selected set based on intrinsic importance and a base method}
\usage{
get_base_set(
  test_statistics,
  p_values,
  alpha = 0.05,
  method = "maxT",
  B = 10000,
  Sigma = diag(1, nrow = length(test_statistics)),
  q = NULL
)
}
\arguments{
\item{test_statistics}{the test statistics (used with "maxT")}

\item{p_values}{(used with "minP" or "Holm")}

\item{alpha}{the alpha level}

\item{method}{the method (one of "maxT", "minP", or "Holm")}

\item{B}{the number of resamples (for minP or maxT)}

\item{Sigma}{the estimated covariance matrix for the test statistics}

\item{q}{the false discovery rate (for method = "BY")}
}
\value{
the initial selected set
}
\description{
Using the estimated intrinsic importance and a base method
designed to control the family-wise error rate (e.g., Holm),
obtain an initial selected set.
}
