% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_spvims.R
\name{pool_spvims}
\alias{pool_spvims}
\title{Pool SPVIM Estimates Using Rubin's Rules}
\usage{
pool_spvims(spvim_ests)
}
\arguments{
\item{spvim_ests}{a list of estimated SPVIMs (of class \code{vim})}
}
\value{
a list of results
}
\description{
If multiple imputation was used due to the presence of missing data,
pool SPVIM estimates from individual imputed datasets using Rubin's rules.
Results in point estimates averaged over the imputations, along with
within-imputation variance estimates and across-imputation variance estimates;
and test statistics and p-values for hypothesis testing.
}
