% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_importance_SL.R
\name{extract_importance_SL}
\alias{extract_importance_SL}
\title{Extract extrinsic importance from a Super Learner object}
\usage{
extract_importance_SL(fit, feature_names, import_type = "all")
}
\arguments{
\item{fit}{the fitted Super Learner ensemble}

\item{feature_names}{the names of the features}

\item{import_type}{the level of granularity for importance: \code{"all"} is the
importance based on the weighted average of ranks across algorithmrithms
(weights are SL coefs); \code{"best"} is the importance based on the algorithmrithm
with highest weight. Defaults to \code{"all"}.}
}
\value{
a tibble, with columns \code{feature} (the feature) and
\code{rank} (the weighted feature importance rank, with 1 indicating the
most important feature).
}
\description{
Extract the individual-algorithm extrinsic importance from each fitted
algorithm within the Super Learner; compute the average weighted rank of the
importance scores, with weights specified by each algorithm's weight in the
Super Learner.
}
