% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_importance_SL_learner.R
\name{extract_importance_SL_learner}
\alias{extract_importance_SL_learner}
\title{Extract the learner-specific importance from a fitted SuperLearner algorithm}
\usage{
extract_importance_SL_learner(fit, coef, feature_names)
}
\arguments{
\item{fit}{the specific learner (e.g., from the Super Learner's
\code{fitLibrary} list).}

\item{coef}{the Super Learner coefficient associated with the learner.}

\item{feature_names}{the feature names}
}
\value{
a tibble, with columns \code{algorithm} (the fitted algorithm),
\code{feature} (the feature), \code{importance} (the algorithm-specific
extrinsic importance of the feature), \code{rank} (the feature importance
rank, with 1 indicating the most important feature), and \code{weight}
(the algorithm's weight in the Super Learner)
}
\description{
Extract the individual-algorithm extrinsic importance from one fitted
algorithm within the Super Learner, along with the importance rank.
}
