% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_augmented_set.R
\name{get_augmented_set}
\alias{get_augmented_set}
\title{Get an augmented set based on the next-most significant variables}
\usage{
get_augmented_set(
  p_values = NULL,
  num_rejected = 0,
  alpha = 0.05,
  quantity = "gFWER",
  q = 0.05,
  k = 1
)
}
\arguments{
\item{p_values}{the adjusted p-values.}

\item{num_rejected}{the number of rejected null hypotheses from the base
FWER-controlling procedure.}

\item{alpha}{the significance level.}

\item{quantity}{the quantity to control (i.e., \code{"gFWER"}, \code{"PFP"},
or \code{"FDR"}).}

\item{q}{the proportion for FDR or PFP control.}

\item{k}{the number of false positives for gFWER control.}
}
\value{
a list of the variables selected into the augmentation set.
}
\description{
Based on the adjusted p-values from a FWER-controlling procedure and a
more general error rate for which control is desired (e.g., generalized
FWER, proportion of false positives, or FDR), augment the set based on FWER
control with the next-most significant variables.
}
