% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrinsic_selection.R
\name{intrinsic_selection}
\alias{intrinsic_selection}
\title{Perform intrinsic, ensemble-based variable selection}
\usage{
intrinsic_selection(
  spvim_ests,
  sample_size,
  alpha = 0.05,
  control = list(quantity = "gFWER", base_method = "Holm", fdr_method = NULL, q = NULL,
    k = NULL)
)
}
\arguments{
\item{spvim_ests}{the estimated SPVIM values (an object of class \code{vim},
resulting from a call to \code{vimp::sp_vim}).}

\item{sample_size}{the number of independent observations used to estimate
the SPVIM values.}

\item{alpha}{the nominal generalized family-wise error rate, proportion of
false positives, or false discovery rate level to control at (e.g., 0.05).}

\item{control}{a list of parameters to control the variable selection process.
Parameters include \code{quantity}, \code{base_method}, \code{q}, and
\code{k}. See \code{\link[flevr]{intrinsic_control}} for details.}
}
\value{
a tibble with the estimated intrinsic variable importance,
the corresponding variable importance ranks, and the selected
variables.
}
\description{
Based on estimated SPVIM values, do variable selection using the
specified error-controlling method.
}
\seealso{
\code{\link[vimp]{sp_vim}} for specific usage of
the \code{sp_vim} function and the \code{vimp} package for estimating
intrinsic variable importance.
}
