% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.ranger.imp.R
\name{SL.ranger.imp}
\alias{SL.ranger.imp}
\title{Super Learner wrapper for a ranger object with variable importance}
\usage{
SL.ranger.imp(
  Y,
  X,
  newX,
  family,
  obsWeights = rep(1, length(Y)),
  num.trees = 500,
  mtry = floor(sqrt(ncol(X))),
  write.forest = TRUE,
  probability = family$family == "binomial",
  min.node.size = ifelse(family$family == "gaussian", 5, 1),
  replace = TRUE,
  sample.fraction = ifelse(replace, 1, 0.632),
  num.threads = 1,
  verbose = FALSE,
  ...
)
}
\description{
Super Learner wrapper for a ranger object with variable importance
}
