% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spvim_vcov.R
\name{spvim_vcov}
\alias{spvim_vcov}
\title{Extract a Variance-Covariance Matrix for SPVIM Estimates}
\usage{
spvim_vcov(spvim_ests)
}
\arguments{
\item{spvim_ests}{estimated SPVIMs}
}
\value{
a variance-covariance matrix
}
\description{
Extract a variance-covariance matrix based on the efficient influence function
for each of the estimated SPVIMs.
}
